+++ packages/camera/camera/test/camera_preview_test.dart
@@ -150,6 +150,9 @@ class FakeController extends ValueNotifier<CameraValue>
   getSupportedVideoStabilizationModes() async => <VideoStabilizationMode>[];
 
   @override
+  Future<void> setJpegImageQuality(int quality) async {}
+
+  @override
   bool supportsImageStreaming() => true;
 }