+++ packages/camera/camera_android_camerax/test/preview_rotation_test.dart
@@ -123,6 +123,7 @@ void main() {
           int? targetRotation,
           CameraXFlashMode? flashMode,
           ResolutionSelector? resolutionSelector,
+          int? jpegQuality,
         }) => MockImageCapture();
     PigeonOverrides.recorder_new =
         ({